requirements.txt

requests
google-generativeai
python-dotenv
streamlit
pandas
youtube-search-python
